backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic-settings==2.0.3
aiofiles==23.2.1
Pillow==10.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
